+++ docs/src/main/paradox/release-notes.md
@@ -5,7 +5,8 @@
 ### 0.8.1
 
 * Added: `rf_rgb_composite` and `rf_render_png`.
-* Added `toMarkdown` and `toHTML` extension methods for `DataFrame`, and registered them with the IPython formatter system when `rf_ipython` is imported.  
+* Added `toMarkdown` and `toHTML` extension methods for `DataFrame`, and registered them with the IPython formatter system when `rf_ipython` is imported.
+* New documentation theme (thanks [@jonas](https://github.com/jonas)!).
 * Fixed: Removed false return type garauntee in cases where an `Expression` accepts either `Tile` or `ProjectedRasterTile` [(#295)](https://github.com/locationtech/rasterframes/issues/295)
 
 ### 0.8.0